@sinhrks 2016-07-22
Updated the bench.``algorithms.algorithm.time_int_unique_duplicated`` now benchmark under ``is_unique`` is once cached.